+++ NEWS.md
@@ -1,3 +1,8 @@
-# cssparser 0.1.0
+# cssparser 0.1.1   2021-10-14
 
-* Initial release  2021-09-29
+* Slightly better `xml_duplicate()` to deal with 'svg'
+
+
+# cssparser 0.1.0   2021-09-29
+
+* Initial release